requirements.txt
torch
transformers
datasets
peft
accelerate
scikit-learn
pyyaml